PD #74901: fix new HLS some bugs
1). new HLS calculate wrong duration.
2). change buffer time value type from int to int64_t.

+++ third_parts/libvhls/hls_main/hls_m3ulivesession.c
@@ -98,7 +98,7 @@ typedef struct _M3ULiveSession{
     int seekflag;
     int codec_data_time;
     int estimate_bandwidth_bps;
-    int cached_data_timeUs;
+    int64_t cached_data_timeUs;
     int is_ts_media;
     int is_encrypt_media;
     int is_http_ignore_range;
